Rename custom slugify into slugs to avoid conflicts
