+++ repondeur/stubs/slugify/__init__.pyi
@@ -1,4 +1,3 @@
 from typing import List, Optional
 
-
 def slugify(text: str, stopwords: Optional[List[str]] = None) -> str: ...
